+++ gifify.sh
@@ -7,7 +7,6 @@ function printHelpAndExit {
   echo 'Options: (all optional)'
   echo '  c CROP:   The x and y crops, from the top left of the image, i.e. 640:480'
   echo '  o OUTPUT: The basename of the file to be output (default "output")'
-  echo '  n:        Do not upload the resulting image to CloudApp'
   echo '  r FPS:    Output at this (frame)rate (default 10)'
   echo '  s SPEED:  Output using this speed modifier (default 1)'
   echo '            NOTE: GIFs max out at 100fps depending on platform. For consistency,'
@@ -20,22 +19,19 @@ function printHelpAndExit {
   exit $1
 }
 
-noupload=0
 fps=10
 speed=1
 
 OPTERR=0
 
-while getopts "c:o:p:r:s:nx" opt; do
+while getopts "c:o:p:r:s:" opt; do
   case $opt in
     c) crop=$OPTARG;;
     h) printHelpAndExit 0;;
     o) output=$OPTARG;;
-    n) noupload=1;;
     p) scale=$OPTARG;;
     r) fps=$OPTARG;;
     s) speed=$OPTARG;;
-    x) cleanup=1;;
     *) printHelpAndExit 1;;
   esac
 done
@@ -83,14 +79,4 @@ ffmpeg -loglevel panic -i "$filename" $filter -r $fps -f image2pipe -vcodec ppm 
 
 echo 'Making gif...'
 cat $temp | convert +dither -layers Optimize -delay $delay - "${output}.gif"
-
-if [ $noupload -ne 1 ]; then
-  open -a CloudApp "${output}.gif"
-
-  if [ $cleanup ]; then
-    rm "$filename"
-    rm "${output}.gif"
-  fi
-else
-  echo "${output}.gif"
-fi
+echo "${output}.gif"